Add extended date validation
Content collection now checks whether the modified date is greater than
or equal to the uploaded date and throws an error if not. Layouts that
pull blogpost date data now display the modified date only if it differs
from the upload date. As a result, modified date is no longer an
optional frontmatter field.

Blog posts have been updated to comply with the schema update.